requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
dj-tasks==1.0.0
django-bootstrap-form==3.4
django-model-utils==4.0.0
django==3.1.2
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1; python_version >= '3.5'
unipath==1.1
